[SW-0] accel/habanalabs: release user interfaces earlier in device fini
Currently the sysfs/debugfs interfaces and device unregistration are
done as the last thing in hl_device_fini(), after several finalizations
and releases are done.
While a disabled flag is set at the beginning of hl_device_fini(), and
it is being checked when handling user accesses to these interfaces,
this check is not hermetic and it better to reverse the order.

Change-Id: Ia5330da9b8696b23480d5782ad4c3cee4b5e719f
Signed-off-by: Tomer Tayar <ttayar@habana.ai>