[SW-123246] hl_sni: relocate ASIC headers
It was decided not to share headers between the compute and the SNI
drivers. Sharing will complicate some copy scripts and anyway in future FS
ASICs, we'll have different repositories so working with copies of the
same header file will be the default mode.
Hence move the headers from the general include folder to the SNI driver.

Change-Id: Ie2321ec3f2dcdd566607a396a3285bb90c3543ce
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>