[SW-0] accel/habanalabs: print qman data on error only for lower qman
In default mode the upper QMANs are not used, and instead engines ARCs
access the lower QMANs directly.
Errors for upper QMANs are therefore not expected, and the debug print
of the PQ entries is not needed.

Modify the QMAN debug data print on errors to include only information
for the lower QMAN.

Change-Id: I52e5186d37b3093b63bdbc8c9545670a05385388
Signed-off-by: Tomer Tayar <ttayar@habana.ai>