[SW-0] accel/habanalabs: update pending reset flags with new reset requests
If hl_device_cond_reset() is called while a reset is already pending but
hasn't started, the reset request will be dropped.
If the flags of the new request are more severe, e.g. a hard reset while
the pending reset is a compute reset, the eventual reset won't be
suitable for the device status.

To prevent such cases, update the pending reset flags with the new
requests flags before the requests are dropped.

Change-Id: I71cabe91a7403f73e2802bd875ffa83b5e268861
Signed-off-by: Tomer Tayar <ttayar@habana.ai>