[SW-110161] gaudi3: verify no NIC on-the-fly AXI transactions after NIC stop
The verification of no on-the-fly AXI transactions for NIC was skipped
until now, because hl_nic_stop() had to be called in a later point.

Allegedly it would have been possible to split hl_nic_stop() into H/W
and S/W parts, and move the H/W parts to be before checking the
transactions counters, but this split is probably too complicated.

Instead, split the counters verification into compute and nic, and
place the nic verification at that later point after hl_nic_stop() is
called.

Change-Id: Icbeeeb5e137c46a1c8f4c1f293f1f3f05e4289e7
Signed-off-by: Tomer Tayar <ttayar@habana.ai>