[SW-114047] accel/habanalabs: add missing cleanup in error path of hl_accel_init
If alloc_chrdev_region() in hl_accel_init() fails, the accel class and
the accel debugfs folder that were previously created are not destroyed.

As part of fixing the error path in hl_accel_init(), modify the
operations order in the function as a preparation to subsequent changes
of registration to the accel subsystem.
With these changes, compiling out the "fake" accel sysfs/debugfs will
require a single ifdef instead of 2.

***** THIS CHANGE SHOULDN'T BE UPSTREAMED *****

Change-Id: I1788ad947c7755990a884d5ff35229e32200fe70
Signed-off-by: Tomer Tayar <ttayar@habana.ai>