[SW-136181] accel/habanalabs: Allow only one registration request at a time
In order to prevent issues while handling re-use of the same ts
offset, we'll allow only one timestamp registration request to be handled
at a time.
Since the flow is protected only by the interrupt lock, and re-use
might request to move the ts node from one interrupt list to another
interrupt, and in such scenario we could end up adding the same node twice
to the interrupts wait lists.

Change-Id: I3109c1dc550f7598b2abee9604b1745f8963a118
Signed-off-by: farah kassabri <fkassabri@habana.ai>