[SW-143566] gaudi3: use interrupt map table for AGG_HDCORE interrupts
Interrupts from blocks on die1 are connected in a reversed order to the
cpu hdcore interrupt aggregators.
In addition, the TPC/CS/RRTR blocks have an "internal" inversion of
instances connectivity for HD 2,3,6,7.
Moreover, the decoder blocks have a "special" connectivity and mapping.

To avoid adding a dedicated mapping in interrupts handler of each hdcore
block, import the F/W interrupt map which is generated automatically
from the excel file, and use it for remapping in the hdcore aggregator
handlers, before calling the blocks specific handlers.

Change-Id: I46c0a0efb1f6dc3d6596762784d5e1b97838af8b
Signed-off-by: Tomer Tayar <ttayar@habana.ai>