[SW-143566] gaudi3: revise prints in interrupts aggregators handlers
Prints in the CPU interrupts aggregators have a
"<block><interrupt_number>" format, which is confusing because it can be
understood as "<block><block_number>".
Moreover, in the CPU hdcore interrupts aggregators, there isn't always a
match between the hdcore number of the aggregator and the hdcore of the
blocks that are connected to it.
Modify the prints such that it would be clearer that the printed number
is the interrupt index, and use the hdcore only for the aggregator name.

While there is no such an issue in the PARC interrupts aggregators
handler, modify it as well to have a similar print pattern.

Change-Id: I562e74b35048c56be05d979f7755d91b85e68047
Signed-off-by: Tomer Tayar <ttayar@habana.ai>