[SW-145099] gaudi3/sim: move priv security config to the head of hw init
To make this possible, we first need to de-isolate all HW blocks to allow
access to all protection bit registers (glbl priv).
Note that this is true only for the case of simulator, since in the case
of an ASIC, privielegd protection bits are configured in a relatively
late stage, way after isolation has been configured.

Change-Id: Ic91a92dcf369a2ca281e458a006c9d6d84d0c93f
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>