[SW-145873] gaudi3: use threaded IRQ handler for interrupt aggregators
While interrupt aggregators messages are disabled in halt_engines, it is
possible that handling of some previous interrupt is still on going.
Synchronizing irqs should have allegedly waited until such a handling
is done, but it just synchronizes with the irq handler itself, and the
actual handling is done in a work of the EQ workqueue, and it can keep
running while we proceed to device reset.

Modify the aggregator interrupts handling to be done in threaded IRQ
handler instead of a work, so synchronizing irqs will indeed wait until
interrupt handling is done.

Change-Id: I89820a9fa817704ae611e89fd30e56e8b51f8156
Signed-off-by: Tomer Tayar <ttayar@habana.ai>