[SW-143622] accel/habanalabs: fix wait_for_interrupt abortion flow
When driver need to abort a waiters for interrupts, for cases
such as critical events that occur and driver need to do hard reset,
in such scenario the driver will complete the fence to wake up the
waiting thread, and will set the fence error indication.
the return value of the completion API will be greater than 0
since it will return the timeout, but driver treat this as wait
finished successfully and it should mark it as aborted.

Change-Id: Ic0e03cfbb31dff4f85bf6896d27983e83ab883c2
Signed-off-by: farah kassabri <fkassabri@habana.ai>