[SW-146097] gaudi3: bugfix in special blocks skip function
When there's a need to access a special block, this function is used
to skip accessing if it isn't accessible for some reason.
One way to skip a special block is if its corresponding ASIC mask
indicates it does not exist, however there was no handling for that
scenario, that is, we handled a case where a block exist, but not
the other way around.

Change-Id: I92dd741b6de897f604f56457e69e221d1bca76fb
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>