[SW-0] accel/habanalabs: fix standalone preboot descriptor request
The preboot used to statically allocate memory for the comms descriptor
on the device memory when driver requested the descriptor information.
Now preboot moved to dynamic memory allocation where it needs the driver
to tell it how much memory to allocate for the comms descriptor.
Note there are no backward compatibility issues as this flow is not
used in production.

Change-Id: Ia9e49f9b76e7140e668a8a4da87ba9065dea5415
Signed-off-by: farah kassabri <fkassabri@habana.ai>