[SW-0] gaudi3: modify gaudi3_handle_eqe() to return rc instead of error count
gaudi3_handle_msg_event() currently handles only NIC_STS, and its
handler function returns 0 for success and not an error count.
In addition, error count isn't relevant also for future MSG events that
will be added to gaudi3_handle_msg_event(), because the reason for MSG
event should be clear and there is no H/W data to read and verify the
presence of an error.
Therefore this function can be modified to return rc instead of an error
count, and for symmetry, modify also gaudi3_handle_hw_event() and handle
the error count internally.

Change-Id: Ia8a08a53fe577553440e89eecfcf970041f2d0d8
Signed-off-by: Tomer Tayar <ttayar@habana.ai>