[SW-147054] accel/habanalabs: set default device release watchdog T/O as 30 sec
After being notified about certain errors, user is expected to finish
his post-errors actions and to release the device within some timeout,
after which is deice is being reset.
The default timeout value is 5 sec, which in some case is not enough for
a user application to collect debug data.
Increase the default value to 30 sec.

Change-Id: I85cce80a34070bbeb28ef8eadaf1affa2196b30c
Signed-off-by: Tomer Tayar <ttayar@habana.ai>