[SW-147940] accel/habanalabs: un-secure PSOC scratchpad register for engine cores interrupt
The F/W dynamically allocates one of the PSOC scratchpad registers for
the engine cores, so they can raise events towards the F/W.
To allow the engine cores to access this register, this register must be
non-secured.

As the driver un-seucres now a register in the PSOC_GLOBAL_CONF block,
the range registers protection (when F/W security is disabled) should be
modified, such that it will skip the PSOC_EFUSE and the PSOC_GLOBAL_CONF
blocks.
These 2 blocks have a PB protection in any case, so this range registers
change shouldn't have an impact.

Change-Id: I5be7a0f1d57e0d59db1263327f776e91d9937848
Signed-off-by: Tomer Tayar <ttayar@habana.ai>