[SW-148743] hl_sni: use dedicated compat file
As a preparation for the upcoming NIC repository split, we need to remove
the dependency on the compute driver.
This patch adds dedicated compat files so we won't use the compute files.
The SNI files were copied from the compute driver and all pure NIC compat
code was removed from the compute side.

Change-Id: I372561df2b795597bbfce83dfb4c827e6e885400
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>