[SW-148743] sni: remove symbols dependency
As a preparation for the upcoming NIC repository split, we need to remove
any dependency between the drivers.
This patch switches the SNI driver handling (in case there is no aux bus)
to use dynamic loading of its probe/remove symbols, so we can remove the
symbols dependency during compute driver build.

Change-Id: I56b36a074387dd4c292ce1a9b9ea597411e39475
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>