[SW-149960] accel/habanalabs: get module refcount
Recently we switched to probe/remove the CN module while using dynamic
lookup of its symbols. In this method the CN module refcount is not
incremented automatically and so the module might be unloaded while the
accel driver is using it.
Hence we must manually increment the module refcount to avoid that.

Change-Id: I56855e6004fd37c0590d992fa70547d744fc1fc5
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>