[SW-122718] accel/habanalabs/gaudi2: prepare to remove soft_rst_irq
The soft reset has transitioned to CPUCP packet instead of plain
register write and is about to be removed from the struct cpu_dyn_regs.
As a preparation for removing the gic_host_soft_rst_irq field from
struct cpu_dyn_regs, switch to use the plain macro - this keeps the
backward compatibility.

Change-Id: I5dd8677d62ba3722463e2c91344776d3668a1c64
Signed-off-by: Igor Grinberg <igrinberg@habana.ai>