[SW-0] accel/habanalabs/gaudi3: ensure valid EDMA/QM error status is passed to iterator
The values of EDMA_CHN_ERR_STATUS and QMAN_GLBL_ERR_STS are received
from H/W upon events, and they are passed to the error cause iterator
together with an array of error strings.
The length of this array is less than 32, and because we can't trust on
the value which is received from F/W, we need to omit any invalid set
bits before calling the iterator.

Change-Id: If7213a60dad83594bf3554f46ab018013328cdf4
Signed-off-by: Tomer Tayar <ttayar@habana.ai>