[SW-152188] accel/habanalabs: split user interrupts pending list
Currently driver maintain one list for both pending user interrupts
which seeks to wait till CQ reaches it's target value and also the ones
that seeks to get timestamp records when the CQ reaches it's target
value.
This causes delay in handling the waiters which gets higher priority
than the timestamp records.
In order to solve this matter driver will split the list into two,
one for each case and each one is protected by it's own spinlock.
Waiters will be handled within the interrupt context first,
then the timestamp records will be set.
freeing the timestamp related memory will be handled in a workqueue.

Change-Id: I16e6b20fd8d111a3f881a8d465c690ce5b1c7533
Signed-off-by: farah kassabri <fkassabri@habana.ai>