[SW-152688] accel/habanalabs/gaudi3: modify event mask and reset flags for NIC interrupts
A reset following an error in the NIC blocks should be a hard reset and
not a compute reset.

In addition, some NIC SPI interrupts are considered as errors and some
are not.
This distinction should be done according to the actual interrupt data.
Temporarily until it is added to the NIC SPI handler, avoid sending
events and reset the device for NIC SPI interrupts.

Change-Id: I37b0775ab1b0c9b08c8a297eb14e4b11459e5ed7
Signed-off-by: Tomer Tayar <ttayar@habana.ai>