[SW-151293] accel/hababalabs/gaudi3: temporarily don't notify user about NIC SEI
In ARC bootrom there is no consideration of which HW blocks exist, which
leads to bad accesses and interrupts on partial PLDM images.
One of these "unreal" interrupts is NIC SEI that involves a notification
to userspace and a hard reset.
Temporarily, until the aforementioned interrupts are cleared on PLDM,
disable the notification and hard reset for NIC SEI.

Change-Id: I0c32c3842035634bf66d345995813ee855f8fc62
Signed-off-by: Tomer Tayar <ttayar@habana.ai>