[SW-154563] accel/habanalabs/gaudi3: increase PLDM sleep time for hard reset
On PLDM there are sometimes AXI drain indications due to accessing
CPU_BOOT_STATUS.
This register is read on hard reset with some delay after initiating the
hard reset command.
Increase the value of this delay as an experiment, to check if the
drain is just a result of a too short delay.

Change-Id: I112cd096a56c113a9041919f0574eae72768a7eb
Signed-off-by: Tomer Tayar <ttayar@habana.ai>