[SW-147637] accel/habanalabs/gaudi2: handle eq health heartbeat check
Add mechanism for eq health check. this will be done using two flows:
using the heartbeat mechanism and raising a dedicated interrupt to
indicate an eq failure like EQ full.
This patch will add implementation for the eq heartbeat for gaudi2 asic.

More info about the heartbeat mechanism:
Expand the heartbeat mechanism to monitor a new event that
will be sent from FW upon receiving heartbeat message.
that way driver can know that the eq is working or not.

Change-Id: Ib78fcdec5965a5d1b276e3d436f4be20d8f6b78d
Signed-off-by: farah kassabri <fkassabri@habana.ai>