[SW-152177] accel/habanalabs/gaudi3: make eq_dynamic_entry dynamic
eq_dynamic_entry is currently defined as a local variable in the
routines that use it when handling the PLDM MSIX interrupts when
no firmware is present (i.e. when no FW is available to handle and
report these interrupts)
These routines are located in gaudi3_bringup.c file
With the above implementation when PSOC eq_dynamic entry was added to
the dynamic eq struct it caused the routines that defined it to exceed
the stack frame limit imposed by the compiler causing compilation
failures.
In order to overcome this issue the EQ entry was removed from being
local to the routine that uses it and is now global to the ISR.
with this commit each PLDM MSIX interrupt is assigned with a dynamic
eq entry at interrupt registration time, and a pointer tos this entry
will be passed to the routines time thus saving stack space.

Change-Id: I1ffedb499774d27628cff13e1197f684f36b8a82
Signed-off-by: Moti Haimovski <mhaimovski@habana.ai>