[SW-124775] accel/habanalabs: set hl_dmabuf_priv.device_address only when needed
The device_address member of 'struct hl_dmabuf_priv' is used only when
virtual device memory is not supported and dma-buf is exported from
address.
Set the value of this field only when it is relevant, and add "phys" to
its name so it would be clearer that it can't be a device virtual
address.

Change-Id: Idf9cc72d666a2f01e12fe0de5d54931dc16caa57
Signed-off-by: Tomer Tayar <ttayar@habana.ai>