[SW-124775] accel/habanalabs: add missing offset handling for dma-buf
On devices with virtual device memory (Gaudi2 onwards), user can provide
an offset within an allocated device memory from which he wants to
export a dma-buf object.
The offset value is verified by driver, but it is not taken into
consideration when the importer driver maps the dma-buf and the SG table
it prepared.
Add the missing offset handling.

Change-Id: Ia2dc9fc3aa7fd4404dea76b51b921e358a5b2465
Signed-off-by: Tomer Tayar <ttayar@habana.ai>