[SW-147637] accel/habanalabs/gaudi2: add eq health check using irq
This is the second patch for applying the eq health check mechanism
which will add support for the interrupt flow for gaudi2 asic.

More info about the interrupt mechanism:
set a dedicated msix for the eq error interrupt, and add
interrupt handler for it.
when FW detects some issue with EQ like EQ_FULL, it'll
raise that interrupt and driver should reset the device.
Driver will inform the FW which msix index to use through
the already existing handshake mechanism which will
send msi info message to fw.

Change-Id: Iaa07290acdd92c08f4223de7db8af486e088ca51
Signed-off-by: farah kassabri <fkassabri@habana.ai>