[SW-124775] accel/habanalabs/compat: limit dma-buf chunk size in old kernel versions
In old kernel versions there is an overflow issue in the ib block
iterator when using large dma-buf sizes.
The fix for this issue didn't involve a change in a header file, which
makes it difficult to detect if we have the fix as part of the compat
mechanism.
Instead, use a direct check of the kernel version, and limit the dma-buf
chunk size if older than the vanilla kernel to which the fix was added.

Change-Id: Ie346f7005dfde393097f1021caf3ed3f98cb6b33
Signed-off-by: Tomer Tayar <ttayar@habana.ai>