[SW-158746] accel/habanalabs: change log level for RMA err
FW set the device unusable indication upon every RMA
error they detect, and the host driver is printing
it as debug log and they dont see it since in most of
systems the driver is not loaded in debug mode.
To help FW debug RMA failures on certain devices
change the log level of this error to be info level.

Change-Id: Ie2c2ea29e0865bad228ecbc6d1b572364ad4afbb
Signed-off-by: farah kassabri <fkassabri@habana.ai>