[SW-160091] accel/habanalabs/gaudi2: perform hard-reset upon PCIe AXI drain event
Non-completed transactions from PCIe towards the device are handled by
the AXI drain mechanism. This handling is in the PCIe level, but the
transactions are still there in the device consuming some queues
entries, and therefore the device must be reset.
Modify to perform hard-reset upon PCIe AXI drain events.

Change-Id: I698144ee72b3689eda8db2301e7a2c61eb901097
Signed-off-by: Tomer Tayar <ttayar@habana.ai>