[SW-152188] accel/habanalabs: fix bug in decoder wait cs
The decoder interrupts are handled in the interrupt context
same as all user interrupts.
so in such case the wait list should be protected by
spin_lock_irqsave in order to avoid deadlock that might happen
with the user submission flow.
this patch change the spin_lock in the decoder wait cs
submission, to be spin_lock_irqsave.

Change-Id: Ie3d0069622f02896d7fe669c37f7be0600a3ab1b
Signed-off-by: farah kassabri <fkassabri@habana.ai>