[SW-160941] accel/habanalabs/cn: ASIC specific headers
Currently some ASIC specific shared structures are embedded in the common
header file cn_aux.h.
This file should be shared also by the Xe driver (as the CN driver is used
for both) but the Gaudi code is irrelevant there.
Moreover, the current approach dictates that the FS1 specific structures
should be added to this file as well which means that FS1 code will be
added to the habanalabs repo which is illogical.
The solution is to move out the ASIC specific code to dedicated header
files (gaudi*_aux.h) so the FS1 code will reside in fs1_aux.h which
will exist only in the Xe repo.

In order to pass CI, this change is done in 3 steps:
1. Move the ASIC specific structures to dedicated headers but still
   include them from cn_aux.h.
2. Once habanalabs is promoted, include the new headers from nic_kmd repo.
3. Once nic_kmd is promoted, remove the include of the new headers from
   cn_aux.h.

This patch is step 1.

Change-Id: Ia7ddc0d765a449981377609b28d772116732547e
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>