[SW-150124] accel/habanalabs/gaudi3: flush pcie queue
For some emulations we see soft lockup due to burst of writes to
registers, and this sometimes kills the pcie link.
As a W/A for the issue, add read after every bunch of writes in order to
flush the queue.
As this patch is updating an auto generated C file, we had to update
the script that generates it and align the excel to the current
configurations we have.
Also add the same flush to the protection bits iterator as
it causes the same issue.

Change-Id: I8b1c140bd0acc2e662b4669799f52c2d38747c6a
Signed-off-by: farah kassabri <fkassabri@habana.ai>