[SW-162577] accel/habanalabs/gaudi3: add support to new MAC OUI
A new OUI is now allowed, hence add support to it.
Also, add support to allow device init even when the OUI is not
recognized, so that the production line won't be blocked.
Hence, treat the check as a sanity check only and print a warning in
case of mismatch.

Change-Id: I464e101949c0ae59196856dd705729df34d63c49
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>