[SW-163643] habanalabs/accel: fix EQ heartbeat mechanism
Use the bfe flag used in heartbeat mechanism to control
whether to perform reset upon EQ check failure or not also.
also don't continue to reschedule of another heartbeat code
when EQ heartbeat check fails as it generates a confusing
logs in dmesg that the heartbeat fails, but it failed actually
because the device state has changed due to the schedule
of hard reset from EQ check.

Change-Id: I0a7cd36699a0e2de99e58e195cf69ba85990796f
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>