[SW-158613] accel/habanalabs/gaudi3: fix pdma eng sei error clear
As clearing the error status register will lead to clear the
STOP-ON-ERR of the engine, this might lead to make the engine
continue with the descriptor even if the current transaction caused
an error.
the correct way to is to mask the engine specific error in the
interrupt error mask register.

Change-Id: I1d1e2413c7d798b93960fd630fcaaa9588bc4336
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>