[SW-162507] accel/habanalabs/gaudi2: move HMMU page tables to HBM
Currently the HMMU page tables are in the host memory,
due to a HW bug in gaudi2 which occurs when HMMU miss happens
and try to fetch the hops from the host along with some other flow
which cause deadlock in HW.
This patch will move HMMU page tables to HBM that way the miss
transaction will read the hops from HBM instead of going to the host.
In addition, for Gaudi2 we'll move the EDMA PQs to HBM to allow
some tests to pass when we block EDMA access to PCIe.

Change-Id: I2bf0a5a736d85bb70e5d1ced6e9ffb73560aee9c
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>