[SW-164631] accel/habanalabs/cn: refactor CN API
The cn_aux.h file must not include the accel header. This is not
upstreamable because it requires the nic_kmd drivers to have a dependency
on it.
This also requires to pass a void pointer for the Gaudi1 map/unmap
functions as the pointer type is not known to nic_kmd (it uses an
identical struct but with a different name).

Change-Id: I27576a21a849f81fc8f09c3d78b784d56799b3b8
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>