[SW-160212] accel/habanalabs/gaudi2: add missing error handling in gaudi2_ctx_init()
gaudi2_ctx_init() lacks a cleanup of hl_cn_ctx_init() in case of a
failure later in the function.
This leads to failing any following context init, because it is marked
in the aux device that the context is already initialized.

Add the missing call to hl_cn_ctx_fini() in the error handling of
gaudi2_ctx_init().

Change-Id: I4a661bf80fbd98b210a1ff1aac45695b768eceb4
Signed-off-by: Tomer Tayar <ttayar@habana.ai>