[SW-0] accel/habanalabs/gaudi3: increase verbosity level of hw_event_info print
Currently the prints inside the interrupts handlers don't include the
info in the EQ aggregator header, like hdcore and instance numbers.
Until we unify the prints for a single event, need to replace dev_dbg()
in gaudi3_print_hw_event_info() with dev_err(), so we will have it in
the kernel log even debug logs are not enabled.

Change-Id: I66931bc47a39f12d6831b200d0b8abb3af0c547b
Signed-off-by: Tomer Tayar <ttayar@habana.ai>