[SW-0] accel/habanalabs/gaudi3: small restructure of EQ events handling
Move the glbl_err handling into a dedicated function, just like it is
for razwi, to make the main SEI handling function cleaner and to ease
the detection of a missing component.

Moreover, move the boot fit existence check to be inside the razwi and
glbl_err functions, as this "common" SEI handling is going to be
extended with at least the mstr_if razwi data handling, and it would be
better to keep the main SEI function simpler.

In addition, in all handling functions, sort the component type switch
case in an alphabetical order, again to make it easier to detect which
handing for which component is missing.

Change-Id: Ifd252a1d45ed71eeef435349dce6796c10ce0541
Signed-off-by: Tomer Tayar <ttayar@habana.ai>