[SW-167875] accel/habanalabs/gaudi3: enable TPC DBG_CFG when enabling debug clock
The TPC debug clock was enabled to allow trace block programming, but
clearing only the DBG_CLK_OFF bit in the TPC_CLK_EN register leads to a
failure later in the TPC QM test.
This problem disappears when clearing also the DBG_CFG_DIS bit, so add
this configuration until the most suitable configuration is clarified.

Change-Id: I0e8e51d4339d5008d4fb5bef016287b69a33d41f
Signed-off-by: Tomer Tayar <ttayar@habana.ai>