[SIV-3058] accel/habanalabs/gaudi3: perform TPC unit reset after soft/hard reset
The TPC QMs are sometimes not idle after a reset or their test fails.
As a temporary workaround, perform a TPC unit reset after a soft/hard
reset.

The unit reset is done after the FW already did its part in the TPC
configuration, so reconfigure it from the driver.

Change-Id: Ib99ea031da77bfbfad789110e4cdfea7776b1247
Signed-off-by: Tomer Tayar <ttayar@habana.ai>