[SW-0] accel/habanalabs/gaudi3: set default masks for engines and binning
The masks for TPC/DEC/ROT are received from FW, so if the communication
with FW is disabled, these masks will be left as zero.

Add a setting of default values to the masks in
gaudi3_set_fixed_properties().
The values will be overridden with the actual masks from FW if the
communication with it is enabled.

Change-Id: If8c499932fa13da61a4df724af5a477c80aeb820
Signed-off-by: Tomer Tayar <ttayar@habana.ai>