[SW-169796] accel/habanalabs/gaudi3/sim: configure privileged RR even if privileged security disabled
Currently hl_init_pb_security() cannot be run in a single-die mode, and
therefore the 'priv_security_enable' flag is not set for it.
As a result, gaudi3_init_security_privileged() is not being run, and
driver doesn't disable the 14th LBW and 8th HBW privileged RR, which
leads to an assertion in the simulator.
Until a support for running hl_iterate_special_blocks() in a single-die
mode is added, run gaudi3_init_lbw_hbw_range_registers_privileged()
without checking the 'priv_security_enable' flag.

Change-Id: Idc60f71c8876dfb75f7f41ff7a553da6d23b412a
Signed-off-by: Tomer Tayar <ttayar@habana.ai>