[SW-0] accel/habanalabs/gaudi3: prevent error cause iterator from exceeding cause array bounds
gaudi3_err_cause_iterator() receives a u32 variable with the error mask
and a pointer to a table with strings that describe the error bits.
In some cases, there are less than 32 valid error bits, and accordingly
the error table has less than 32 strings.
If in these cases the error mask from FW includes set bits which are
above the max expected error bit, the iterator will exceed the table
bounds.

Provide the iterator also the table size as a parameter, to detect these
cases and trim the error mask from FW if needed.

Change-Id: Ifd67195a93b564921c8f9f40c0644f025b77438f
Signed-off-by: Tomer Tayar <ttayar@habana.ai>