[SW-169652] accel/habanalabs/gaudi3: get pll data using cpucp packet
During Gaudi3 development we needed to get the PLL frequency even
without FW.

To support this a helper function was added to obtain the data using
registers access.

Now that FW is the main flow we should obtain this data using standard
CPUCP packet.

The old functionality is still left for PLDM case.

Change-Id: If420ea1ce8ab0de445ec53e55ec1be26230ec1c2
Signed-off-by: Ohad Sharabi <osharabi@habana.ai>