[FSW-2697] accel/habanalabs/gaudi2: remove aux func can_unset_asid_cfg()
During integration with the FW this code had to align with the version
to distinguish between FW versions that support this feature and
those who don't.

From a backward compatibility point of view, we no longer have to
maintain compatibility to those old FW versions so we can remove them.

Now that this function is no longer in use in NIC's code, we can remove
the callback from LKD code.

Change-Id: I8e9ef2d29f838a5e1124c0e4fef66650b1add089
Signed-off-by: Ohad Sharabi <osharabi@habana.ai>