[SW-0] accel/habanalabs: modify print for skip loading linux FW to debug log
Skip loading a linux FW image into the device with the current supported
ASICs is done for test purposes only.
Moreover, for future supported ASICs it is possible that there won't be
a need to load such an image.
The print in such a case is therefore not needed in most cases, so
replace the used dev_info() with dev_dbg().

Change-Id: Iec5546587a165157ad5f19476ec84b96f605496a
Signed-off-by: Tomer Tayar <ttayar@habana.ai>