[SW-170064] accel/habanalabs/cn: move gaudi{1,2,3} aux ops to device specific files
As a process of decoupling the NIC FS1 driver from the Habanalabs
driver, removing any aux ops that are not needed by FS1 from the generic
aux ops structure.

This patch adds the aux ops to the device specific structures and
provides the implementation.

The functions will be removed from the generic structure in a follow-up
patch.

Change-Id: I08c74c8e2d6adcf318dec95239ae510f0a26eac3
Signed-off-by: Tal Risin <trisin@habana.ai>