[SW-173086] accel/habanalabs: keep explicit size of reserved memory for FW
The reserved memory for FW is currently saved in an ASIC property in
units of MB, just like the value that comes from FW.
Except the fact that it is not clear from the property's name, it means
also that a calculation to actual size is required everywhere that it is
used.
Modify the property to hold the size in bytes.

Change-Id: Ibd0066f56c15a39caa433dc8cb3bb92b4f90fb72
Signed-off-by: Tomer Tayar <ttayar@habana.ai>