[SW-172371] accel/habanalabs/gaudi3: temporarily force HBW drain timeout configuration
The estimated max HBW latency is ~2500 nsec, and if taking a margin of
x10, it is ~25,000 nsec.
The relevant clock is 1GHz (25,000 = 0x61A8 clock cycles), so modify the
currently configured value of 0x1000 clock cycles with 0x8000.

For now temporarily force this configuration also when working with FW,
until the updated values are configured there.

Change-Id: Icb19dec63360004ab6f6c2b1e3ee594384c72291
Signed-off-by: Tomer Tayar <ttayar@habana.ai>