[SW-173086] accel/habanalabs/gaudi3: statically reserve first 32MB of HBM to FW
The addresses of the reserved HBM memories are calculated once in
set_fixed_properties(), and later again in hw_init() to shift the
addresses if an optional memory reservation is requested by FW.
One of these reserved addresses is used for the HMMU page table, which
is used in mmu_init() for creating a genpool. This genpool is created
before hw_init() is called, and there is no simple way to update the
genpool's chunk address if needed.
Instead of updating the genpool in a less acceptable method, or changing
the order of the init operations, modify the HBM allocation for FW to be
a static reservation of 32MB.

Change-Id: Idd5635457b0798966d1dfb430a6c5702c0ccbd13
Signed-off-by: Tomer Tayar <ttayar@habana.ai>