[SW-173707] accel/habanalabs/gaudi3: fix a deadlock in hard reset upon an EQ event
The EQ handling for Gaudi3 is done with a threaded IRQ and not with the
EQ workqueue that was used for previous devices.
When a hard reset is requested from the IRQ thread, the EQ IRQ
synchronization which follows the disable PCI access request from FW,
leads to a deadlock.

As it appears that a threaded IRQ is not suitable in this case, modify
the handling to use the EQ workqueue.
However, as opposed to previous devices, no work objects are allocated
dynamically in the IRQ handler and the EQ entries are not copied to such
objects, as in Gaudi3 a much larger EQ entry size is used.

Change-Id: I5fa411b457d7607e69f0466ed8f511394ea9fc5c
Signed-off-by: Tomer Tayar <ttayar@habana.ai>