[SW-174581] accel/habanalabs: modify pci health check
Today we read PCI VENDOR-ID in order to make sure PCI link is
healthy. Apparently the VENDOR-ID might be stored on host and
hence, when we read it we might not access the PCI bus.
In order to make sure PCI health check is reliable, we will start
checking the DEVICE-ID instead.

Change-Id: I5169c524471c68a8b842822c0dea5971346e2139
Signed-off-by: Ofir Bitton <obitton@habana.ai>