[SW-1754832] accel/habanalabs: add heartbeat debug info
It's kind of hard to debug the reason for heartbeat check failure.
As attempt to ease this task, this patch will provide more information
when this failure happens, in order to help understand what's wrong.
As this heartbeat check the communication with FW, so printing
the CPU queue pi/ci and the counter of how many times that event
was received would help in debugging the issue.

Change-Id: I2ea4aafc5b32882893fd03931e467db632c263e7
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>