[SW-175482] accel/habanalabs/gaudi2/3: set heartbeat debug info
It's kind of hard to debug the reason for heartbeat check failure.
As attempt to ease this task, this patch will provide more information
when this failure happens, in order to help understand what's wrong.
As this heartbeat check the communication with FW, so printing
the counter of how many times that event was received would
help in debugging the issue.
In addition due to change in new releases FW boot flow,
it takes longer to process the first heartbeat message,
so we'll wait longer before sending the first heartbeat message.

Change-Id: Ie97526d8fa06d89ba488b7df7c12fe8f0e5707fc
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>