[SW-173701] accel/habanalabs: check for errors after preboot is ready
LKD should check and report any fatal errors detected by preboot,
before it attempts to load the boot fit.
Some errors may cause the driver to stop the boot process and mark
the device as unusable.
This check will allow the driver to fail and print the error reported
by preboot and skip the time wasting attempt of trying to load the
boot fit, which will fail due to the error.

Change-Id: I80364b9de1b5b3a61549c2b068ae209def606323
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>