[SW-0] accel/habanalabs: add module params to extend bringup flexibility
When using the module param ignore_fw_nic_info, we ignore FW info
and therefore don't have any configuration for serdes type and for card
location. This prevents the flexibility to manipulate the HW configuration
and testing at the bringup stage, when full flexibility is required
to test all scenarios.

Change-Id: Ibde26321e23da459286bbcf363f668c23ada22fc
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>