[SW-176799] accel/habanalabs/gaudi3: send unmask_irq request for engines EQ events
Currently if any kind of a reset should be performed when receiving an
EQ event, the unmask_irq request is not sent to FW.
If this reset is a compute reset, FW will keep the relevant interrupt
masked after the compute reset is done.
Modify to not send the unmask_irq request only for events that require a
hard reset.

As a compute reset can escalate to a hard reset in
hl_device_cond_reset(), it is not always needed to send the unmask_irq
request when the intended reset is a compute reset.
To avoid such redundant requests, call hl_fw_unmask_irq() only after
hl_device_cond_reset(), so if a hard reset was eventually scheduled, the
disabled flag will be set and the request will be dropped.

Change-Id: I4289843ab3e04d3a9d4e21e85c213612b99657fc
Signed-off-by: Tomer Tayar <ttayar@habana.ai>