[SW-176635] accel/habanalabs/gaudi3: use a suitable EQ size for EQ init/fini/reset
The EQ init/fini/reset functions use the default HL_EQ_SIZE_IN_BYTES
value, which is not suitable for Gaudi3 because it uses the dynamic EQ
mechanism.
As a result, not enough memory is allocated from the CPU-accessible DMA
pool, and later allocations (e.g. for cpucp packets) override the
content of the EQ.

Update the relevant ASIC property with the suitable value.

Change-Id: Ia7f1f3583029b60c4fa210e9b94c9c8a1f0b583c
Signed-off-by: Tomer Tayar <ttayar@habana.ai>