[SW-180109] - makefile: don't run kernel-doc script check on some kernels
For versions 5.10.210 and up (all until 5.11), there is a bug in
kernel-doc script. This script checks for non-conventional documentation,
and warns about it. Since we treat warnings as errors, a warning there
will fail the compilation.

As there is an inherent bug in these versions, the script always fails,
regardless to the actual documentation.

In order to workaround this issue, for these versions only, skip
the kernel-doc script. The kernel-doc script is triggered once W=1 in the
makefile. But, W=1 also holds more warning flags. Therefore, for the
versions above, add all warning flags which would have been received by
setting W=1, but do not set W=1. This way, all checks will be run, expect
the kernel-doc one, and only for the versions with the bug.

Change-Id: Ie4b9ba3398f86adc90aa5ea2728f342fe9e0c1ce
Signed-off-by: David Meriin <dmeriin@habana.ai>