[SW-181200] accel/habanalabs/gaudi3: update ticket number in a TODO comment
SW-175925 is currently the ticket number in the TODO comment to stop
disabling the MME/TPC regulators.
This is an embedded ticket, and even when it is done we will need to
wait for the FW deployment before removing this code.
Replace it with another ticket that is assigned to the driver team.

Change-Id: I92e21e457acc5c94484805e1bb5db19e9d9da05e
Signed-off-by: Tomer Tayar <ttayar@habana.ai>