[SW-180918] accel/habanalabs/gaudi3: don't check dropped commands as part of CS invalidation
Currently as part of the invalidation of the CS blocks, it is verified
that there are no ongoing maintenance/privileged commands, and also that
there are no dropped commands.
After modifying the handling of the dropped commands to not reset the
device, it is possible to have such commands upon soft reset, and
therefore it should not be verified as part of the invalidation.

Change-Id: I4a6d8f3c23c15218bd79a9310702cdb27828dafb
Signed-off-by: Tomer Tayar <ttayar@habana.ai>