[SW-0] accel/habanalabs/gaudi3: revise the print upon CS invalidation failure
The address of the CACHE_MAINT_STATUS is currently printed upon a CS
invalidation failure, probably to indicate in which CS the failure was.
However gaudi3_cs_invalidation_poll_status() is called from the CS
iterator, and therefore the information of HD/CS number is already
printed upon failure.
As printing this address is redundant, remove it and use instead the
last read value of the register.

Change-Id: I23bc3268a283a4c1c7e5358d6256e0fef6f424e9
Signed-off-by: Tomer Tayar <ttayar@habana.ai>