[SW-181182] accel/habanalabs: change the heartbeat scheduling point
Currently we schedule the heartbeat thread at late init,
cn init and only then we set FW the INTS_REGISTER
packet to enable events to be received from them.
cn init takes around 10 seconds, which steals from
the time we wanted to give to FW to get ready to respond
to heartbeat packets. As we wanted to give FW full 2 cycles
of the heartbeat thread scheduling to get ready after receiving
INTS_REGISTER, and only then start checking whether we received
respond or not.
The patch will move the heartbeat thread scheduling to be after driver
is done with all initializations.

Change-Id: I97c7ef8da15bd030456594e514a620c2847b4f48
Signed-off-by: Farah Kassabri <fkassabri@habana.ai>