[SW-182329] accel/habanalabs: revise print on EQ heartbeat failure
Don't print the "previous EQ index" value in case of a EQ heartbeat
failure, because it is incremented along with the EQ CI and therefore
redundant.

In addition, as the CPU-CP PI is zeroed when it reaches a value that is
twice the queue size, add a value of the CI with a similar wrap around,
to make it easier to compare the values.

Change-Id: I61198309e7e3d7b2eb1acc0b54248d5e9f762883
Signed-off-by: Tomer Tayar <ttayar@habana.ai>