[SW-159630] accel/habanalabs/cn: abandon aux bus ctx_init/fini
In Gaudi1-3 drivers, the compute driver used to inform the CN driver on
every new user context.
Today, in Gaudi2 and above (including FS family), the NIC relevant user
contexts will be ibverbs user context and will be passed to the CN driver
by the IB driver.
Therefore, the old infra between the compute and the CN driver regarding
new ctx is needed only for Gaudi1 and the common aux bus functions can
be removed.

Change-Id: Ib68349164bce09f2345cb475e38e888e726dd1a6
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>