[SW-0] makefile: add + to GIT_SHA str if there are local changes
Currently, one can not know from the dmesg logs whether the driver was
loaded while there are local changes, it will show the last git commit
in either case.

To be able to see if the user added some local changes we will add a +
suffix to the driver GIT_SHA as already the standard in Linux kernel
versions.

For example, the version string will be:
without local changes: I45e1fcf
with local changes: I45e1fcf+

Change-Id: I45e1fcffbea8a19bc80fac2a7c33e5a221c97752
Signed-off-by: Ohad Sharabi <osharabi@habana.ai>