[SW-181592] accel/habanalabs/gaudi3: reenable drv-security only if FW security is off
In general, the driver can configure both levels of security,
i.e., secured & privileged, both of which are independent.
When it comes to privileged, we have no problems, because it
can only be enabled if there's no FW loaded, or the FW loaded
is unsecured (that way, there can't be potential conflicts).
But when it comes to secured-level security, it becomes a bit
tricky now that a secured-FW enables ISEC security by default,
and before the driver is loaded, which means that driver access
(via PCI) to certain PBs will be restricted.
Due to the complexity of handling these conflicts, it was
decided to fully disable driver-level security when a secured
FW is loaded.

Change-Id: Ib848cd164dd4922782d7ed177c8f90f11fd7e903
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>