[SW-186005] accel/habanalabs/gaudi3: don't use the hard_reset_done_by_fw ASIC property
In case of a failure in FW that prevents it from updating the
FW_HARD_RST_EN bit, the "hard_reset_done_by_fw" ASIC property will be
left unset, and in the following call to hl_device_fini(), the driver
will try to hard reset the device by its own.

This attempt might fail, and the relevant error message will be
confusing, because this code path is not expected.

As in Gaudi3 there is no case of working with FW and resetting from
driver, stop using this attribute and instead use the FW components
mask.
This will prevent the attempt to reset by driver even if the FW
capability bit is not set.

Change-Id: Ifbb0819b6b3f45004a58416eac047f5904a3037e
Signed-off-by: Tomer Tayar <ttayar@habana.ai>