[SW-0] accel/habanalabs: return ENOMEM if less than requested pages were pinned
EFAULT is currently returned if less than requested user pages are
pinned.
This value means a "bad address" which might be confusing to user, as
the address of the given user memory is not necessarily "bad".

Modify the return value in to ENOMEM, as "out of memory" is more
suitable in this case.

Change-Id: I0b010a3c21c8b902136e46445a955ea8a3fd7a24
Signed-off-by: Tomer Tayar <ttayar@habana.ai>